requirements.txt
# Python dependencies
kafka-python
psycopg2-binary
python-dotenv
pandas
streamlit
pytest
plotly
